+++ cmd/commands.go
@@ -153,6 +153,7 @@ var (
 		&appsInfoCommand,
 		&openCommand,
 		&dashboardCommand,
+		&appsProjectSetCommand,
 
 		// Apps Actions
 		&logsCommand,